refactor: configloader package refactor
