+++ internal/pkg/pluginengine/cmd_apply.go
@@ -12,7 +12,7 @@ import (
 )
 
 func Apply(configFile string, continueDirectly bool) error {
-	cfg, err := configloader.LoadConf(configFile)
+	cfg, err := configloader.LoadConfig(configFile)
 	if err != nil {
 		return err
 	}
