+++ internal/pkg/show/status/status.go
@@ -27,7 +27,7 @@ func Show(configFile string) error {
 		id = "default"
 	}
 
-	cfg, err := configloader.LoadConf(configFile)
+	cfg, err := configloader.LoadConfig(configFile)
 	if err != nil {
 		return err
 	}
